Fix types problem
unsigned int was used instead of unsigned long, and intmax instead of
uintmax

+++ includes/variadic_args_defs.h
@@ -6,7 +6,7 @@
 /*   By: mgautier <mgautier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/02/22 10:32:31 by mgautier          #+#    #+#             */
-/*   Updated: 2017/03/13 16:47:42 by mgautier         ###   ########.fr       */
+/*   Updated: 2017/03/16 18:27:35 by mgautier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -55,28 +55,28 @@ struct						s_var_arg
 {
 	union
 	{
-		int					t_int;
-		long int			t_long_int;
-		long long int		t_long_long_int;
-		intmax_t			t_intmax_t;
-		ptrdiff_t			t_ptrdiff_t;
-		ssize_t				t_ssize_t;
-		unsigned int		t_uint;
-		unsigned int		t_ulong_int;
-		unsigned long int	t_ulong_long_int;
-		uintmax_t			t_uintmax_t;
-		size_t				t_size_t;
-		char				*t_ptr_char;
-		short				*t_ptr_short;
-		int					*t_ptr_int;
-		long int			*t_ptr_long_int;
-		long long int		*t_ptr_long_long_int;
-		intmax_t			*t_ptr_intmax_t;
-		ptrdiff_t			*t_ptr_ptrdiff_t;
-		size_t				*t_ptr_size_t;
-		wint_t				t_wint_t;
-		wchar_t				*t_wstring;
-	}						parameter;
+		int						t_int;
+		long int				t_long_int;
+		long long int			t_long_long_int;
+		intmax_t				t_intmax_t;
+		ptrdiff_t				t_ptrdiff_t;
+		ssize_t					t_ssize_t;
+		unsigned int			t_uint;
+		unsigned long int		t_ulong_int;
+		unsigned long long int	t_ulong_long_int;
+		uintmax_t				t_uintmax_t;
+		size_t					t_size_t;
+		char					*t_ptr_char;
+		short					*t_ptr_short;
+		int						*t_ptr_int;
+		long int				*t_ptr_long_int;
+		long long int			*t_ptr_long_long_int;
+		intmax_t				*t_ptr_intmax_t;
+		ptrdiff_t				*t_ptr_ptrdiff_t;
+		size_t					*t_ptr_size_t;
+		wint_t					t_wint_t;
+		wchar_t					*t_wstring;
+	}							parameter;
 	t_type					type;
 };
